Fix CI
