+++ scorecard/tests/test_scorecard.py
@@ -1,6 +1,11 @@
 from django.contrib.sites.models import Site
 from django.core.management import call_command
 
+from django.test import (
+    TransactionTestCase,
+    Client,
+)
+
 from municipal_finance.tests.helpers import BaseSeleniumTestCase
 
 
@@ -35,3 +40,20 @@ class ScorecardTest(BaseSeleniumTestCase):
         #selenium.get("%s%s" % (self.live_server_url, "/profiles/municipality-BUF-buffalo-city/"))
         #is_displayed = selenium.find_element_by_css_selector('.is--pre-2019-20').is_displayed()
         #assert is_displayed == False
+
+
+class GeographyDetailViewTestCase(TransactionTestCase):
+    serialized_rollback = True
+
+    def test_context(self):
+        # Make request
+        client = Client()
+        response = client.get("/profiles/municipality-BUF-buffalo-city/")
+        context = response.context
+        page_data = json.loads(context["page_data_json"])
+        # Test for amount types
+        self.assertIsInstance(page_data["amount_types_v1"], dict)
+        # Test for cube names
+        self.assertIsInstance(page_data["cube_names"], dict)
+        # Test for municipality category descriptions
+        self.assertIsInstance(page_data["municipal_category_descriptions"], dict)
